# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LA GLORIA STORE API",
        "description": "THE OFFICIAL LA GLORIA F.C. STORE",
        "contact": {
            "email": "dieztomasjoaquin@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/rest/brands": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get all brands",
                "description": "Returns a paginated list of all brands",
                "operationId": "6497956451bb6ed13ec486c1576f8253",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BrandResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/rest/brands/{id}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Get a brand by ID",
                "description": "Returns a single brand by its ID",
                "operationId": "7428292740aad6cb46add1fca717cbb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the brand",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            }
        },
        "/rest/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Returns a paginated list of all categories",
                "operationId": "6123f24a834bf0cdb03f669352d73bcc",
                "responses": {
                    "200": {
                        "description": "List of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No categories found"
                    }
                }
            }
        },
        "/rest/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category by ID",
                "description": "Returns a single category by ID",
                "operationId": "61609f8e2e4c5e96ff6ec1fcca428059",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
        "/rest/order-details": {
            "get": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Get a list of all order details",
                "description": "Returns a paginated list of all order details in the application",
                "operationId": "65c1a0b9ac70c18ba1cf5235a0d2e19a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDetailResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rest/order-details/{id}": {
            "get": {
                "tags": [
                    "Order Details"
                ],
                "summary": "Get a single order detail",
                "description": "Returns the order detail identified by the given ID",
                "operationId": "getOrderDetailById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order detail to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetailResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BrandResource": {
                "title": "Brand resource",
                "description": "Represents a Brand",
                "properties": {
                    "id": {
                        "description": "The brand's ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The brand's name",
                        "type": "string",
                        "example": "Nike"
                    },
                    "created_at": {
                        "description": "The datetime when the brand was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-11 13:42:56"
                    },
                    "updated_at": {
                        "description": "The datetime when the brand was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-11 13:42:56"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "title": "Category Resource",
                "description": "Represents a category",
                "properties": {
                    "id": {
                        "description": "The ID of the category",
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "description": "The name of the category",
                        "type": "string",
                        "example": "Shoes"
                    },
                    "created_at": {
                        "description": "The date and time when the category was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the category was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "OrderDetailResource": {
                "title": "Order Detail Resource",
                "description": "Represents an Order detail",
                "properties": {
                    "id": {
                        "description": "The ID of the order detail",
                        "type": "integer",
                        "example": 1
                    },
                    "product_amount": {
                        "description": "The amount of the product in the order detail",
                        "type": "integer",
                        "example": 3
                    },
                    "product": {
                        "$ref": "#/components/schemas/ProductResource"
                    },
                    "created_at": {
                        "description": "The date and time when the order detail was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-12 14:30:00"
                    },
                    "updated_at": {
                        "description": "The date and time when the order detail was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-05-12 14:32:00"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "ProductResource",
                "description": "Represents a Product",
                "properties": {
                    "id": {
                        "description": "The product's unique identifier",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the product",
                        "type": "string"
                    },
                    "size": {
                        "description": "The size of the product",
                        "type": "string"
                    },
                    "image": {
                        "description": "The URL of the product's image",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the product",
                        "type": "number",
                        "format": "float"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/BrandResource"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "created_at": {
                        "description": "The date and time when the product was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}